Migrate BaseSched into Scheduler
