+++ AnkiDroid/src/test/java/com/ichi2/libanki/sched/SchedulerTest.kt
@@ -875,16 +875,16 @@ open class SchedulerTest : RobolectricTest() {
         c2.load()
         Assert.assertEquals(QUEUE_TYPE_SIBLING_BURIED, c2.queue)
         assertNull(col.sched.card)
-        col.sched.unburyCardsForDeck(BaseSched.UnburyType.MANUAL)
+        col.sched.unburyCardsForDeck(Scheduler.UnburyType.MANUAL)
         c.load()
         Assert.assertEquals(QUEUE_TYPE_NEW, c.queue)
         c2.load()
         Assert.assertEquals(QUEUE_TYPE_SIBLING_BURIED, c2.queue)
-        col.sched.unburyCardsForDeck(BaseSched.UnburyType.SIBLINGS)
+        col.sched.unburyCardsForDeck(Scheduler.UnburyType.SIBLINGS)
         c2.load()
         Assert.assertEquals(QUEUE_TYPE_NEW, c2.queue)
         col.sched.buryCards(listOf(c.id, c2.id))
-        col.sched.unburyCardsForDeck(BaseSched.UnburyType.ALL)
+        col.sched.unburyCardsForDeck(Scheduler.UnburyType.ALL)
         Assert.assertEquals(Counts(2, 0, 0), col.sched.counts())
     }
